Files of the voltimeter project.
The project was developed for NUCLEO-F446RE board using the Atollic TrueSTUDIO for STM32 and Code::Blocks IDE's. The commentary was all written in portuguese, so I will explain how the project works. In the microcontroller, the ADC picks up the voltage on the potentiometer and converts to a eight bit word, and then, sends to the USART2, whose peripheral is linked to the ST-LINK.
The data is received in the COM3(in my case), and then, the SerialCOM program prints a value of 0 to 255 multiplied by the ADC resolution(3.3V/255).